style: rename axios folder
